INFRA-505: Move stuff around
